Removed timestamping, audit() now in base class
Removed timestamp-checking on run-directories. There are other better methods to check the validity of a run-directory.
Audit() method is now a base Job class method, as most of the variations between them were removed.
