+++ sequence_processing_pipeline/FastQCJob.py
@@ -1,5 +1,5 @@
-from os import listdir, makedirs, walk
-from os.path import basename, exists, join
+from os import listdir, makedirs
+from os.path import exists, join
 from sequence_processing_pipeline.Job import Job
 from sequence_processing_pipeline.PipelineError import PipelineError
 from functools import partial
@@ -42,6 +42,7 @@ class FastQCJob(Job):
         # for lists greater than n commands, chain the extra commands,
         # distributing them evenly throughout the first n commands.
         self.commands = self._group_commands(self.commands)
+        self.suffix = 'fastqc.html'
 
         self._generate_job_script()
 
@@ -232,38 +233,3 @@ class FastQCJob(Job):
 
         with open(sh_details_fp, 'w') as f:
             f.write('\n'.join(self.commands))
-
-    def audit(self, sample_ids):
-        '''
-        Audit the results of a run.
-        :param sample_ids: A list of sample-ids that require results.
-        :return: A list of sample-ids that were not found.
-        '''
-        files_found = []
-
-        # assume for now that a corresponding zip file exists for each html
-        # file found. Assume for now that all html files will be found in a
-        # 'filtered_sequences' or 'trimmed_sequences' subdirectory.
-        #
-        # verify that the entire list of sample-ids found match what's
-        # expected. Since the list of expected ids is very small, we'll
-        # perform an exact comparison.
-        for root, dirs, files in walk(self.output_path):
-            files_found += [join(root, x) for x in files if
-                            x.endswith('fastqc.html')]
-
-        # remove all files found with a 'zero_files' directory in the path
-        files_found = [x for x in files_found if 'zero_files' not in x]
-
-        found = []
-        for sample_id in sample_ids:
-            for found_file in files_found:
-                # the trailing underscore is important as it can be assumed
-                # that all fastq.gz files will begin with sample_id followed
-                # by an '_', and then one or more additional parameters
-                # separated by underscores. This substring is unlikely to be
-                if basename(found_file).startswith('%s_' % sample_id):
-                    found.append(sample_id)
-                    break
-
-        return sorted(list(set(found) ^ set(sample_ids)))
