+++ sequence_processing_pipeline/configuration.json
@@ -1,8 +1,6 @@
 {
   "configuration": {
     "pipeline": {
-      "younger_than": 168,
-      "older_than": 24,
       "archive_path": "sequence_processing_pipeline/tests/data/sequencing/knight_lab_completed_runs",
       "search_paths": [ "/tmp", "sequence_processing_pipeline/tests/data"]
     },
