+++ sequence_processing_pipeline/tests/test_Pipeline.py
@@ -3,10 +3,8 @@ import os
 from sequence_processing_pipeline.PipelineError import PipelineError
 from sequence_processing_pipeline.Pipeline import Pipeline
 import unittest
-from os import utime, makedirs
+from os import makedirs
 from os.path import abspath, basename, join
-from copy import deepcopy
-from time import time
 from functools import partial
 import re
 
@@ -42,24 +40,6 @@ class TestPipeline(unittest.TestCase):
         # most of the test code within 'with open()' expressions.
         with open(self.good_config_file, 'r') as f:
             self.good_config = json.load(f)
-
-        # create a fake age for self.good_run_dir by taking the current time
-        # and subtracting from it a value in seconds that's in the middle of
-        # the allowable age-range.
-        config = self.good_config['configuration']['pipeline']
-        valid_age = ((config['younger_than'] - config[
-                      'older_than']) / 2) * 3600
-
-        self.new_timestamp = time() - valid_age
-
-    def reset_run_directory_timestamp(self, timestamp=None):
-        '''
-        Reset the timestamp of self.good_run_dir.
-        :param timestamp: Reset to supplied value or known good value if None.
-        :return:
-        '''
-        ts = timestamp if timestamp is not None else self.new_timestamp
-        utime(self.good_run_dir, (ts, ts))
 
     def tearDown(self):
         # Pipeline is now the only class aware of these files, hence they
@@ -100,10 +80,6 @@ class TestPipeline(unittest.TestCase):
         # Pipeline object will raise an Error.
         self.delete_runinfo_file()
 
-        # reset the timestamp on self.good_run_dir because it was just
-        # created.
-        self.reset_run_directory_timestamp()
-
         with self.assertRaisesRegex(PipelineError, "required file 'RunInfo.xml"
                                                    "' is not present."):
             Pipeline(self.good_config_file, self.good_run_id,
@@ -114,9 +90,6 @@ class TestPipeline(unittest.TestCase):
         # an Error is raised when only RTAComplete.txt is missing.
         self.delete_rtacomplete_file()
         self.create_runinfo_file()
-
-        # with the above changes we need to adjust the timestamp again.
-        self.reset_run_directory_timestamp()
 
         with self.assertRaisesRegex(PipelineError, "required file 'RTAComplete"
                                                    ".txt' is not present."):
@@ -129,9 +102,6 @@ class TestPipeline(unittest.TestCase):
         self.create_rtacomplete_file()
         self.make_runinfo_file_unreadable()
 
-        # with the above changes we need to adjust the timestamp again.
-        self.reset_run_directory_timestamp()
-
         with self.assertRaisesRegex(PipelineError, "RunInfo.xml is present, bu"
                                                    "t not readable"):
             Pipeline(self.good_config_file, self.good_run_id,
@@ -140,10 +110,6 @@ class TestPipeline(unittest.TestCase):
         self.make_runinfo_file_readable()
 
     def test_creation(self):
-        # reset the timestamp on self.good_run_dir because it was just
-        # created.
-        self.reset_run_directory_timestamp()
-
         # Pipeline should assert due to config_file
         with self.assertRaises(PipelineError) as e:
             Pipeline(self.bad_config_file,
@@ -205,109 +171,7 @@ class TestPipeline(unittest.TestCase):
                      self.good_qiita_id,
                      None)
 
-        with open(join('sequence_processing_pipeline', 'configuration.json'),
-                  'r') as f:
-            cfg = json.load(f)
-            with self.assertRaises(PipelineError) as e:
-                cpy_cfg = deepcopy(cfg)
-                cpy_cfg['configuration']['pipeline']['younger_than'] = -1
-                Pipeline(self.good_config_file,
-                         self.good_run_id,
-                         self.good_sample_sheet_path,
-                         self.good_output_file_path,
-                         self.good_qiita_id,
-                         cpy_cfg)
-
-            self.assertEqual(str(e.exception), 'older_than and younger_than '
-                                               'cannot be less than zero.')
-            with self.assertRaises(PipelineError) as e:
-                cpy_cfg = deepcopy(cfg)
-                cpy_cfg['configuration']['pipeline']['older_than'] = -1
-                Pipeline(self.good_config_file,
-                         self.good_run_id,
-                         self.good_sample_sheet_path,
-                         self.good_output_file_path,
-                         self.good_qiita_id,
-                         cpy_cfg)
-            self.assertEqual(str(e.exception), 'older_than and younger_than '
-                                               'cannot be less than zero.')
-            with self.assertRaises(PipelineError) as e:
-                cpy_cfg = deepcopy(cfg)
-                cpy_cfg['configuration']['pipeline']['younger_than'] = 20
-                cpy_cfg['configuration']['pipeline']['older_than'] = 30
-                Pipeline(self.good_config_file,
-                         self.good_run_id,
-                         self.good_sample_sheet_path,
-                         self.good_output_file_path,
-                         self.good_qiita_id,
-                         cpy_cfg)
-            self.assertEqual(str(e.exception), 'older_than cannot be equal to '
-                                               'or less than younger_than.')
-
-    def test_filter_directories_for_time(self):
-        cfg = deepcopy(self.good_config)
-
-        # set a range that 211021_A00000_0000_SAMPLE's timestamps must
-        # fall within: between 1 and 2 hours.
-        cfg['configuration']['pipeline']['younger_than'] = 2
-        cfg['configuration']['pipeline']['older_than'] = 1
-
-        # Set directory's timestamp to between one and two hours and
-        # verify it is returned by filter_directories_for_time().
-
-        # get the current time in seconds since the epoch.
-        current_time = time()
-        # create an epoch time value older than 1 hour ago + 5 min.
-        older_than = current_time - 3900
-        self.reset_run_directory_timestamp(timestamp=older_than)
-
-        try:
-            Pipeline(self.good_config_file, self.good_run_id,
-                     self.good_sample_sheet_path, self.good_output_file_path,
-                     self.good_qiita_id, cfg)
-        except PipelineError as e:
-            self.fail(("test_filter_directories_for_time failed w/PipelineEr"
-                       f"ror: {e.message}"))
-
-        # Set directory's timestamp to just older than two hours and verify it
-        # is not returned by filter_directories_for_time().
-
-        # get the current time in seconds since the epoch.
-        current_time = time()
-        # create an epoch time value older than 2 hour ago + 5 min.
-        older_than = current_time - 7500
-        self.reset_run_directory_timestamp(timestamp=older_than)
-
-        with self.assertRaises(PipelineError) as e:
-            Pipeline(self.good_config_file, self.good_run_id,
-                     self.good_sample_sheet_path, self.good_output_file_path,
-                     self.good_qiita_id, cfg)
-        self.assertEqual(str(e.exception), ('sequence_processing_pipeline/tes'
-                                            'ts/data/211021_A00000_0000_SAMPL'
-                                            'E is too old.'))
-
-        # Set directory's timestamp to just under one hour and
-        # verify it is not returned by filter_directories_for_time().
-
-        # get the current time in seconds since the epoch.
-        current_time = time()
-        # create an epoch time value younger than 1 hour ago.
-        older_than = current_time - 3300
-        self.reset_run_directory_timestamp(timestamp=older_than)
-
-        with self.assertRaises(PipelineError) as e:
-            Pipeline(self.good_config_file, self.good_run_id,
-                     self.good_sample_sheet_path, self.good_output_file_path,
-                     self.good_qiita_id, cfg)
-        self.assertEqual(str(e.exception), ('sequence_processing_pipeline/tes'
-                                            'ts/data/211021_A00000_0000_SAMPL'
-                                            'E is too young.'))
-
     def test_sample_sheet_validation(self):
-        # reset the timestamp on self.good_run_dir because it was just
-        # created.
-        self.reset_run_directory_timestamp()
-
         # test successful validation of a good sample-sheet.
         # if self.good_sample_sheet_path points to a bad sample-sheet, then
         # Pipeline would raise a PipelineError w/warnings and error messages
@@ -330,10 +194,6 @@ class TestPipeline(unittest.TestCase):
                                             'EP479894B04'))
 
     def test_generate_sample_information_files(self):
-        # reset the timestamp on self.good_run_dir because it was just
-        # created.
-        self.reset_run_directory_timestamp()
-
         # test sample-information-file generation.
         pipeline = Pipeline(self.good_config_file, self.good_run_id,
                             self.good_sample_sheet_path,
